Fixed awkward rendering with Julia set and Koch
Also changed render calls to use a super method to avoid cody copy
pasting
Also fixed test cases.
