+++ CS320_Team_Fractal_Website/junit/edu/ycp/cs320/CS320_Team_Fractal_Website/controller/BarnsleyControllerTest.java
@@ -59,10 +59,10 @@ public class BarnsleyControllerTest {
 		params = model.getParameters();
 		assertTrue(controller.acceptParameters(params));
 	}
-	
+
 	@Test
-	public void testRender(){
-		assertTrue(controller.render());
+	public void testRenderImage(){
+		assertFalse(controller.renderImage() == null);
 	}
 	
 	@Test
