+++ CS320_Team_Fractal_Website/junit/edu/ycp/cs320/CS320_Team_Fractal_Website/controller/FractalControllerTest.java
@@ -9,6 +9,7 @@ import org.junit.Test;
 
 import edu.ycp.cs320.CS320_Team_Fractal_Website.controller.fractal.FractalController;
 import edu.ycp.cs320.CS320_Team_Fractal_Website.model.fractal.Fractal;
+import edu.ycp.cs320.CS320_Team_Fractal_Website.model.fractal.Sierpinski;
 
 public class FractalControllerTest{
 	
@@ -18,8 +19,8 @@ public class FractalControllerTest{
 	public void setUp(){
 		controller = new FractalController() {
 			@Override
-			public boolean render() {
-				return true;
+			public BufferedImage renderImage() {
+				return null;
 			}
 			@Override
 			public Fractal getModel() {
@@ -53,6 +54,13 @@ public class FractalControllerTest{
 	public void testSendImage(){
 		BufferedImage img = new BufferedImage(10, 10, BufferedImage.TYPE_3BYTE_BGR);
 		assertTrue(controller.sendImage(img));
+	}
+	
+	@Test
+	public void testRender(){
+		Fractal f = new Sierpinski();
+		FractalController c = f.createApproprateController();
+		assertTrue(c.render());
 	}
 	
 }
