+++ CS320_Team_Fractal_Website/junit/edu/ycp/cs320/CS320_Team_Fractal_Website/controller/SierpinskiControllerTest.java
@@ -57,10 +57,10 @@ public class SierpinskiControllerTest{
 		params = model.getParameters();
 		assertTrue(controller.acceptParameters(params));
 	}
-	
+
 	@Test
-	public void testRender(){
-		assertTrue(controller.render());
+	public void testRenderImage(){
+		assertFalse(controller.renderImage() == null);
 	}
 	
 	@Test
