+++ CS320_Team_Fractal_Website/src/edu/ycp/cs320/CS320_Team_Fractal_Website/controller/fractal/BarnsleyController.java
@@ -54,25 +54,9 @@ public class BarnsleyController extends FractalController{
 	}
 	
 	@Override
-	public boolean render(){
-		//create a thread to render and calculate the set
-		//TODO make this use multiple threads
-		Thread thread = new Thread(new Runnable(){
-			@Override
-			public void run(){
-				BufferedImage img = renderBarnleyFern();
-				sendImage(img);
-			}
-		});
-		
-		thread.start();
-		try{
-			thread.join();
-		}catch (InterruptedException e){
-			e.printStackTrace();
-			return false;
-		}
-		return true;
+	public BufferedImage renderImage(){
+		BufferedImage img = renderBarnleyFern();
+		return img;
 	}
 	
 	/**
