+++ CS320_Team_Fractal_Website/src/edu/ycp/cs320/CS320_Team_Fractal_Website/controller/fractal/KochController.java
@@ -45,38 +45,22 @@ public class KochController extends FractalController{
 	}
 	
 	@Override
-	public boolean render(){
-		//create a thread to render and calculate the set
-		//TODO make this use multiple threads
-		Thread thread = new Thread(new Runnable(){
-			@Override
-			public void run(){
-				BufferedImage img = new BufferedImage(SIZE, SIZE, BufferedImage.TYPE_3BYTE_BGR);
-				Graphics g = img.getGraphics();
-				
-				g.setColor(Color.WHITE);
-				g.fillRect(0, 0, SIZE, SIZE);
-				
-				//size of triangle
-				double s = .9 * SIZE * Math.sqrt(3.0) / 2.0;
-				//y offset
-				double y = -50;
-				renderKochLine(g, new Point2D.Double(s, s + y), new Point2D.Double(SIZE - s, s + y), model.getIterations());
-				renderKochLine(g, new Point2D.Double(SIZE / 2, SIZE - s + y), new Point2D.Double(s, s + y), model.getIterations());
-				renderKochLine(g, new Point2D.Double(SIZE - s, s + y), new Point2D.Double(SIZE / 2, SIZE - s + y), model.getIterations());
-				
-				sendImage(img);
-			}
-		});
+	public BufferedImage renderImage(){
+		BufferedImage img = new BufferedImage(SIZE, SIZE, BufferedImage.TYPE_3BYTE_BGR);
+		Graphics g = img.getGraphics();
 		
-		thread.start();
-		try{
-			thread.join();
-		}catch (InterruptedException e){
-			e.printStackTrace();
-			return false;
-		}
-		return true;
+		g.setColor(Color.WHITE);
+		g.fillRect(0, 0, SIZE, SIZE);
+		
+		//size of triangle
+		double s = .9 * SIZE * Math.sqrt(3.0) / 2.0;
+		//y offset
+		double y = -50;
+		renderKochLine(g, new Point2D.Double(s, s + y), new Point2D.Double(SIZE - s, s + y), model.getIterations());
+		renderKochLine(g, new Point2D.Double(SIZE / 2, SIZE - s + y), new Point2D.Double(s, s + y), model.getIterations());
+		renderKochLine(g, new Point2D.Double(SIZE - s, s + y), new Point2D.Double(SIZE / 2, SIZE - s + y), model.getIterations());
+		
+		return img;
 	}
 	
 	public static void renderKochLine(Graphics g, Point2D.Double start, Point2D.Double end, int iterations){
@@ -90,31 +74,33 @@ public class KochController extends FractalController{
 		Point2D.Double point = getEquilateralPoint(low, high);
 		
 		//draw new lines
-		g.setColor(Color.BLACK);
-		g.drawLine(
-				(int)Math.round(start.x),
-				(int)Math.round(start.y),
-				(int)Math.round(low.x),
-				(int)Math.round(low.y)
-		);
-		g.drawLine(
-				(int)Math.round(end.x),
-				(int)Math.round(end.y),
-				(int)Math.round(high.x),
-				(int)Math.round(high.y)
-		);
-		g.drawLine(
-				(int)Math.round(low.x),
-				(int)Math.round(low.y),
-				(int)Math.round(point.x),
-				(int)Math.round(point.y)
-		);
-		g.drawLine(
-				(int)Math.round(high.x),
-				(int)Math.round(high.y),
-				(int)Math.round(point.x),
-				(int)Math.round(point.y)
-		);
+		if(iterations == 1){
+			g.setColor(Color.BLACK);
+			g.drawLine(
+					(int)Math.round(start.x),
+					(int)Math.round(start.y),
+					(int)Math.round(low.x),
+					(int)Math.round(low.y)
+			);
+			g.drawLine(
+					(int)Math.round(end.x),
+					(int)Math.round(end.y),
+					(int)Math.round(high.x),
+					(int)Math.round(high.y)
+			);
+			g.drawLine(
+					(int)Math.round(low.x),
+					(int)Math.round(low.y),
+					(int)Math.round(point.x),
+					(int)Math.round(point.y)
+			);
+			g.drawLine(
+					(int)Math.round(high.x),
+					(int)Math.round(high.y),
+					(int)Math.round(point.x),
+					(int)Math.round(point.y)
+			);
+		}
 		
 		//call more lines
 		renderKochLine(g, point, high, iterations - 1);
