+++ CS320_Team_Fractal_Website/src/edu/ycp/cs320/CS320_Team_Fractal_Website/servlet/MainPageServlet.java
@@ -213,4 +213,5 @@ public class MainPageServlet extends HttpServlet{
 		//if no username is found, return null
 		return null;
 	}
+	
 }
